+++ database/migrations/2024_04_07_161056_create_kri_penilaians_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('kri_penilaians', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('sub_penilaian_id');
-            $table->integer('bobot');
+            $table->decimal('bobot');
             $table->foreign('sub_penilaian_id')->references('id')->on('sub_penilaians');
             $table->timestamps();
         });
